requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.1.0
attrs==21.4.0
autopep8==1.4.3
billiard==3.6.4.0
braintree==3.14.0
celery==5.2.0
certifi==2019.3.9
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.5.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-appconf==1.0.5
django-celery-beat==2.2.1
django-compressor==3.1
django-compressor-toolkit==0.6.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-discover-runner==1.0
django-environ==0.8.1
django-formtools==2.3
django-paypal==1.1.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-restframework==0.0.1
django-role-permissions==3.1.1
django-timezone-field==4.2.1
django-tinymce4-lite==1.7.4
django.js==0.8.1
djangorestframework==3.12.4
factory-boy==3.0.1
Faker==11.3.0
filelock==3.4.2
fudge==1.1.1
gunicorn==19.9.0
identify==2.4.7
idna==2.8
iniconfig==1.1.1
isort==4.3.4
jsmin==2.2.2
kombu==5.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
nodeenv==1.6.0
oauthlib==3.0.1
packaging==21.3
pep8==1.7.1
phonenumbers==8.12.41
Pillow==7.1.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.5.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.2.2
pyparsing==3.0.7
pytest==6.2.5
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.1.0
pytz==2018.9
PyYAML==6.0
rcssmin==1.1.0
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.2.0
simplejson==3.17.6
six==1.12.0
sqlparse==0.3.0
stripe==2.62.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.24.2
vine==5.0.0
virtualenv==20.13.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.11.0
